Add pairwise best replacement strategy
This is used in Differential Evolution, add bib references and cite
papers/book. Also describe selection pressure of *replacement*
strategies, so we can use a selection scheme without selection pressure
(tournament selection with a tournament size of 1) and a replacement
strategy *with* selection pressure to prevent the search from
degenerating into a random walk.