Fix PGAGetBestIndex with constraints
Now PGAGetBestIndex computes the best index with constraints.
This made problems when using the fitness for comparison of individuals
in PGAStringCompare because the fitness is not updated for certain
replacement strategies. In addition it was discovered that certain badly
scaled problems where subtracting a small evaluation from a large one
during default fitness-scaling will degenerate (and subtract 0).  This
results in PGAGetBestIndex displaying the wrong Individuals.  So
PGAStringCompare now doesn't use the fitness but compares evaluations
directly. This has resulted in one badly-scaled problem to find a better
solution because the default Tournament selection now correctly compares
individuals (it uses PGAStringCompare). It also has resulted in a test
failing that now produces a slightly worse evaluation (after first
getting ahead of the original optimization).