Fix evaluation comparison
The comparison of evaluation was sorting in the wrong direction.
Now the PGAEvalCompare and PGAEvalSort are in utils.c.
Rename the element 'evalfunc' of PGAIndividual to 'evalue': The name
evalfunc suggests a function pointer. The new name needed to be
searchable with 'grep' and reflect something shorter than
'evaluation_value'.
Introduce a back-pointer from PGAIndividual to the context structure.
This enables us to sort an array of PGAIndividual pointers from which the
indeces of Individuals can be recovered without resorting to tricks with
static data structures. Introduce new functions starting with 'IND' that
get at least one Individual (and no Context) as parameters.
Refactor PGASortPop to use the population check once at the beginning.