Implement NSGA-II
The Nondominated Sorting Genetic Algorithm (Version 2) NSGA-II was
described by Deb et.al. 2002, see examples/nsgaii/README.rst.
NSGA-II in PGApack is implemented as a new replacement scheme.
It implements multi-objective optimization.

Rename property of PGAIndividual auxtotaluptodate to auxtotalok.

Rewrite of almost all statistics computation and printing: Offline and
Online metrics should not use evaluations where constraints are not
satisfied. So these now only count unconstrained solutions. We now
compute the best evaluations taking multi-objectives and constraints
into account: For each function the best value is computed separately.

Printing of the worst evaluation is not supported in the presence of
multi-objective optimization. The notion of a worst value is ill-defined
in that context.