Fixes in statistics computation
Formatting: Most of the test changes are whitespace fixes to make the
additional column for constraint violations align.
The c_maxbit_1.data differences are very small rounding errors in the
online performance.
The changes in the deb_optimize_X_1.data are due to changes in how the
statistics are computed: Average, Online, and Offline now only use
individuals which have no constraint violations.
Note that in addition to fixes in the statistics, the computation of the
worst individual was also changed to take constraints into account.