Add serialization/deserialization
This adds a variable-length encoding for genes: The transmitted
serialization need not be the same length each time. This can be done
(even though MPI doesn't allow variable-length data) because we add a
message with the real length in front of the real one. This can be done
because we do not use multicast.