Honor NoDuplicates option for NSGA algorithms
These combine the new and the old generation. So even if we do not
generate duplicates during one generation, by combining two generations
with possibly the same individuals in them, we would generate
duplicates. Now when combining the two generations we sort the
duplicates last, so they do not end up in the new combined generation.

Changes a lot of tests where duplicate checking is enabled with one of
the NSGA variants.