Add index to PGAIndividual
.. and transfer this index via MPI.
This allows us to have the index of the individual in the MPI process
performing the evaluation. Can be useful for submitting a job-request
(e.g. via MPI) to another service: When using evaluation via MPI, the
index of the individual used for evaluation in that instance is the same
(negative) constant in all MPI processes. Note that pgapack does not use
that index.