Replace PGAError* with PGAFatalPrintf
.. where applicable (i.e. only in the fatal error case).
Deprecate PGA_FATAL flag and PGAError. With printf-like error printing
functions we can better check the arguments (instead of passing void
pointers like in PGAError). In addition with a special method that
*always* exits (PGAFatalPrintf) we can now give the compiler more info
that error checks that do not pass actually never return -- so that it
can establish that, e.g., after a malloc we do not use a NULL pointer.