Fix 2nn crowding metric
Add the second closest distance as crowding2 to PGAIndividual.
Change crowdingplot to allow multiple solutions in a single file.
This can be used to plot the output in test/unit/pop_test.data.
Set all crowding metrics of 2nn and mnn to 0 for removed individuals.
This ensures that they will be removed during rank/crowding sorting.
Add tests with 2nn and mnn crowding metrics.
Call zdt4 with popsize 200 (instead of 100) and 500 generations, with
lower popsize it would not always find the best pareto front (and we
want to compare crowding metrics).